About popup - fixed, animations are better Possible to mute sounds and allow/not allow application to open safe cells in zero directions. Score table keeps the best hundred now. Everything out of the line will be removed. Fixed bug when speedmeter arrow jumped (±5 degrees) at the leftmost and rightmost positions.

+++ JapaneseMinesweeper/JMSLeaderboardTableViewController.m
@@ -102,8 +102,9 @@
     cell.lbLevel.text = [gameSession.level stringValue];
     
     NSUInteger progress = lroundf(gameSession.progress.floatValue * 100);
-    cell.lbProgress.text = [NSString stringWithFormat:@"%d%%", progress];
+    cell.lbProgress.text = [NSString stringWithFormat:@"%lu%%", (unsigned long)progress];
     cell.lbProgress.textColor = progress == 100 ? [UIColor colorFromInteger:0xff009900] : [UIColor colorFromInteger:0xffff7f00];
+
     return cell;
 }
